+++ docs/_pages/releases.md
@@ -12,6 +12,10 @@ sidebar:
 ### What's new
 
 * Introduced a new assembly-level attribute that you can use to initialize Fluent Assertions before the first assertion - [#2292](https://github.com/fluentassertions/fluentassertions/pull/2292)
+* Introduced new collection assertions methods - [#2432](https://github.com/fluentassertions/fluentassertions/pull/2432)
+  * `BeProperSubsetOf`
+  * `BeProperSupersetOf`
+  * `BeSupersetOf`
 
 ### Improvements
 * Improve failure message for string assertions when checking for equality - [#2307](https://github.com/fluentassertions/fluentassertions/pull/2307)